adding dark mode toggle
To see how to use dark mode toggle and persist it, see here: https://www.100ms.live/blog/jotai-react-state-management.
Removing the flash is nontrivial.
To get rid of it, see the index.html file.
This is based on the output of Next.js static website generated with the
next-themes package to toggle the theme.
I had to change it slightly, since the Next.js static website included
System as a theme option.
Also, the Jotai theme provider uses JSON.stringify to encode the value,
so I had to use JSON.parse to get it.